backend/requirements.txt
<<<<<<< HEAD
## Minimal, compulsory backend dependencies for development

# Core framework
=======
<<<<<<< HEAD
## Minimal, compulsory backend dependencies for development

# Core framework
Django==5.0.4
djangorestframework==3.14.0

# JWT authentication (project settings reference simplejwt)
djangorestframework-simplejwt

# CORS (needed if frontend runs on separate origin)
django-cors-headers==3.13.0

# Database driver (traceback used PyMySQL)
PyMySQL
mysqlclient==2.2.4

# Channels + WebSockets
=======
# ==============================
# Core Framework
# ==============================
>>>>>>> 4518809402808e6ca5db71f9c83e1773a55a49d8
Django==5.0.4
djangorestframework==3.14.0

# JWT authentication (project settings reference simplejwt)
djangorestframework-simplejwt

# CORS (needed if frontend runs on separate origin)
django-cors-headers==3.13.0

# Database driver (traceback used PyMySQL)
PyMySQL
mysqlclient==2.2.4

<<<<<<< HEAD
# Channels + WebSockets
=======
# ==============================
# WebSockets & Real-time (Channels)
# ==============================
>>>>>>> d4cac4720d4cace9d447997545cd6c0a379eefc5
>>>>>>> 4518809402808e6ca5db71f9c83e1773a55a49d8
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

<<<<<<< HEAD
# Image handling
Pillow==10.3.0

# Environment and HTTP
=======
<<<<<<< HEAD
# Image handling
Pillow==10.3.0

# Environment and HTTP
=======
# ==============================
# Media & Image Handling
# ==============================
Pillow==10.3.0

# ==============================
# Environment & HTTP
# ==============================
>>>>>>> d4cac4720d4cace9d447997545cd6c0a379eefc5
>>>>>>> 4518809402808e6ca5db71f9c83e1773a55a49d8
python-dotenv==1.0.1
requests==2.32.3
